server/requirements.txt
fastapi[standard]
openai
python-dotenv
serpapi
google-search-results
python-multipart
torch
transformers